requirements.txt
attrs==25.3.0
beautifulsoup4>=4.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
exceptiongroup==1.2.2
fake-useragent>=0.1.11
h11==0.14.0
idna==3.10
importlib_resources==6.5.2
numpy>=1.21.0
outcome==1.3.0.post0
packaging==24.2
pandas>=1.5.0
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv>=1.0.0
pytz==2025.2
requests>=2.25.0
selenium>=4.0.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
webdriver-manager>=3.8.0
websocket-client==1.8.0
wsproto==1.2.0
zipp==3.21.0
matplotlib>=3.5.0
seaborn>=0.12.0
flask>=2.0.0
flask-wtf>=1.0.0
plotly>=5.0.0
scikit-learn>=1.0.0
folium>=0.12.0
gunicorn>=20.1.0
